# Allow optimizing locking for built-in exemplar reservoirs 