--- a/sdk/metric/exemplar/fixed_size_reservoir_test.go
+++ b/sdk/metric/exemplar/fixed_size_reservoir_test.go
@@ -45,7 +45,7 @@ func TestNewFixedSizeReservoirSamplingCorrectness(t *testing.T) {
 	}
 
 	var sum float64
-	for _, m := range r.store {
+	for _, m := range r.measurements {
 		sum += m.Value.Float64()
 	}
 	mean := sum / float64(sampleSize)