--- a/sdk/metric/internal/aggregate/filtered_reservoir.go
+++ b/sdk/metric/internal/aggregate/filtered_reservoir.go
@@ -5,10 +5,12 @@ package aggregate // import "go.opentelemetry.io/otel/sdk/metric/internal/aggreg
 
 import (
 	"context"
+	"sync"
 	"time"
 
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/sdk/metric/exemplar"
+	"go.opentelemetry.io/otel/sdk/metric/internal/x"
 )
 
 // FilteredExemplarReservoir wraps a [exemplar.Reservoir] with a filter.
@@ -29,6 +31,11 @@ type FilteredExemplarReservoir[N int64 | float64] interface {
 type filteredExemplarReservoir[N int64 | float64] struct {
 	filter    exemplar.Filter
 	reservoir exemplar.Reservoir
+	// The exemplar.Reservoir is not required to be concurrent safe, but
+	// implementations can indicate that they are concurrent-safe by embedding
+	// x.ConcurrentSafeReservoir in order to improve performance.
+	reservoirMux   sync.Mutex
+	concurrentSafe bool
 }
 
 // NewFilteredExemplarReservoir creates a [FilteredExemplarReservoir] which only offers values
@@ -37,17 +44,30 @@ func NewFilteredExemplarReservoir[N int64 | float64](
 	f exemplar.Filter,
 	r exemplar.Reservoir,
 ) FilteredExemplarReservoir[N] {
+	_, concurrentSafe := r.(x.ConcurrentSafeReservoir)
 	return &filteredExemplarReservoir[N]{
-		filter:    f,
-		reservoir: r,
+		filter:         f,
+		reservoir:      r,
+		concurrentSafe: concurrentSafe,
 	}
 }
 
 func (f *filteredExemplarReservoir[N]) Offer(ctx context.Context, val N, attr []attribute.KeyValue) {
 	if f.filter(ctx) {
+		ts := time.Now()
+		if !f.concurrentSafe {
+			f.reservoirMux.Lock()
+			defer f.reservoirMux.Unlock()
+		}
 		// only record the current time if we are sampling this measurement.
-		f.reservoir.Offer(ctx, time.Now(), exemplar.NewValue(val), attr)
+		f.reservoir.Offer(ctx, ts, exemplar.NewValue(val), attr)
 	}
 }
 
-func (f *filteredExemplarReservoir[N]) Collect(dest *[]exemplar.Exemplar) { f.reservoir.Collect(dest) }
+func (f *filteredExemplarReservoir[N]) Collect(dest *[]exemplar.Exemplar) {
+	if !f.concurrentSafe {
+		f.reservoirMux.Lock()
+		defer f.reservoirMux.Unlock()
+	}
+	f.reservoir.Collect(dest)
+}