--- a/sdk/metric/internal/reservoir/doc.go
+++ b/sdk/metric/internal/reservoir/doc.go
@@ -1,0 +1,6 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+// Package reservoir contains experimental features used by built-in exemplar
+// reservoirs which require coordination with the metrics SDK.
+package reservoir // import "go.opentelemetry.io/otel/sdk/metric/internal/reservoir"